Move klt to external
