+++ external/klt/src/convolve.cc
