+++ external/klt/src/pnmio.cc
