+++ external/klt/src/pyramid.cc
