+++ external/klt/src/storeFeatures.cc
