+++ external/klt/src/writeFeatures.cc
